blogi_backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
python-multipart
python-jose
